Add files via upload
Answering math word problems is a difficult task, and requires logical reasoning over implicit or explicit quantities expressed in text. In this assignment, our
goal is develop an automatic converter, which would convert the textual narrative into executable program. One of the ways to generate executable programs
from their textual description is to make use of Encoder-Decoder architecture. For this part of assignment, we will make use of seq2seq architectures
to convert input text into the corresponding symbolic form. 

The folder contains the train, test and dev json files along with a checker file.
Each json contains a P rogram, linear formula and answer fields for each
example. The train split contains 19791 samples, dev contains 2961 and test
contains 1924 samples. We would be generating the linear formula given the
Problem, the example of same is given in 1.